# perf: improve performance of cnpj algorithm