```
goos: linux
goarch: amd64
pkg: github.com/phenpessoa/br
cpu: Intel(R) Core(TM) Ultra 7 155H
                         │   old.txt   │               new.txt                │
                         │   sec/op    │    sec/op     vs base                │
CNPJ_IsValid18-21          40.70n ± 2%   26.92n ±  3%  -33.85% (p=0.000 n=10)
CNPJ_IsValid14-21          43.59n ± 2%   18.79n ±  1%  -56.91% (p=0.000 n=10)
CNPJ_IsValid18Invalid-21   40.29n ± 1%   26.71n ±  3%  -33.69% (p=0.000 n=10)
CNPJ_IsValid14Invalid-21   43.96n ± 1%   18.64n ±  1%  -57.60% (p=0.000 n=10)
CNPJ_String18-21           61.80n ± 5%   48.02n ± 11%  -22.30% (p=0.000 n=10)
CNPJ_String14-21           98.23n ± 0%   75.81n ±  1%  -22.82% (p=0.000 n=10)
geomean                    51.69n        31.17n        -39.70%
```